+++ source/YugiAcademy/YuGiOh/cards/__init__.py
@@ -1,0 +1,3 @@
+from .monster_card import MonsterCard
+from .spell_card import SpellCard
+from .trap_card import TrapCard
