+++ source/YugiAcademy/YuGiOh/views.py
@@ -1,7 +1,5 @@
 from django.shortcuts import render, redirect
-from cards.monster_card import MonsterCard
-from cards.spell_card import SpellCard
-from cards.trap_card import TrapCard
+from cards import MonsterCard, SpellCard, TrapCard
 from booster_packs.booster_pack import BoosterPack
 from booster_packs.booster_pack_card import BoosterPackCard
 from .forms import SpellCardForm, SearchBoosterPackForm, BoosterPackForm, BoosterPackCardForm
